separando em arquivos as funções da api. > > Co-authored-by: paola-oliveira <paolanatalia_oliveira@outlook.com>
