+++ src/Componentes/PedidosEntregues/PedidosEntregues.js
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import { obterPedidos } from '../../API/api';
+import { obterPedidos } from '../../API/Orders.js';
 import { useEffect, useState } from 'react';
 import Menu from '../Menu/Menu';
