+++ src/__tests__/AdicionarProdutosApi.test.js
@@ -1,4 +1,4 @@
-import { adicionarProdutos } from "../API/api"
+import { adicionarProdutos } from "../API/Products.js"
 import * as localStorageUtil from '../util/localStorage';
 
 describe('adicionarProdutos', () => {
